requirements.txt
Django==2.2.4
django-extensions==2.1.6
pytest==4.4.0
psycopg2==2.8.3
djangorestframework==3.10.3 
Werkzeug==0.16.0
pytest-django==3.6.0
mixer==6.1.3 